+++ src/main/java/com/secuconnect/client/api/PaymentContractsApi.java
@@ -752,4 +752,86 @@ public class PaymentContractsApi {
         apiClient.executeAsync(call, localVarReturnType, callback);
         return call;
     }
+
+    /**
+     * Build call for revokeAccrual
+     * @param paymentContractId Contract identifier (required)
+     * @return Call to execute
+     * @throws ApiException If fail to serialize the request body object
+     */
+    public Call revokeAccrualCall(String paymentContractId) throws ApiException {
+        Object localVarPostBody = null;
+
+        // create path and map variables
+        String localVarPath = "/Payment/Contracts/{paymentContractId}/revokeAccrual"
+            .replaceAll("\\{" + "paymentContractId" + "\\}", apiClient.escapeString(paymentContractId.toString()));
+
+        List<Pair> localVarQueryParams = new ArrayList<Pair>();
+        List<Pair> localVarCollectionQueryParams = new ArrayList<Pair>();
+
+        Map<String, String> localVarHeaderParams = new HashMap<String, String>();
+
+        Map<String, Object> localVarFormParams = new HashMap<String, Object>();
+
+        final String[] localVarAccepts = {
+            "application/json"
+        };
+        final String localVarAccept = apiClient.selectHeaderAccept(localVarAccepts);
+        if (localVarAccept != null) localVarHeaderParams.put("Accept", localVarAccept);
+
+        final String[] localVarContentTypes = {
+            
+        };
+        final String localVarContentType = apiClient.selectHeaderContentType(localVarContentTypes);
+        localVarHeaderParams.put("Content-Type", localVarContentType);
+
+        String[] localVarAuthNames = new String[] { "oauth_token" };
+        return apiClient.buildCall(localVarPath, "POST", localVarQueryParams, localVarCollectionQueryParams, localVarPostBody, localVarHeaderParams, localVarFormParams, localVarAuthNames);
+    }
+
+    @SuppressWarnings("rawtypes")
+    private Call revokeAccrualValidateBeforeCall(String paymentContractId) throws ApiException {
+        // verify the required parameter 'paymentContractId' is set
+        if (paymentContractId == null) {
+            throw new ApiException("Missing the required parameter 'paymentContractId' when calling revokeAccrual(Async)");
+        }
+
+        return revokeAccrualCall(paymentContractId);
+    }
+
+    /**
+     * POST Payment/Contracts/{paymentContractId}/revokeAccrual
+     * Revoke accrual flag for all transactions of given contract
+     * @param paymentContractId Contract identifier (required)
+     * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
+     */
+    public void revokeAccrual(String paymentContractId) throws ApiException {
+        revokeAccrualWithHttpInfo(paymentContractId);
+    }
+
+    /**
+     * POST Payment/Contracts/{paymentContractId}/revokeAccrual
+     * Revoke accrual flag for all transactions of given contract
+     * @param paymentContractId Contract identifier (required)
+     * @return ApiResponse&lt;Void&gt;
+     * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
+     */
+    public ApiResponse<Void> revokeAccrualWithHttpInfo(String paymentContractId) throws ApiException {
+        Call call = revokeAccrualValidateBeforeCall(paymentContractId);
+        return apiClient.execute(call);
+    }
+
+    /**
+     * POST Payment/Contracts/{paymentContractId}/revokeAccrual (asynchronously)
+     * Revoke accrual flag for all transactions of given contract
+     * @param paymentContractId Contract identifier (required)
+     * @param callback The callback to be executed when the API call finishes
+     * @return The request call
+     * @throws ApiException If fail to process the API call, e.g. serializing the request body object
+     */
+    public Call revokeAccrualAsync(String paymentContractId, final ApiCallback<Void> callback) throws ApiException {
+        Call call = revokeAccrualValidateBeforeCall(paymentContractId);
+        apiClient.executeAsync(call, callback);
+        return call;
+    }
 }